requirements.txt
numpy
mediapipe
opencv-python
Pillow
imageio